store wallet by deterministic name rather than checksum, refactor, and cleanup
